test(dataset_broker): use random ports
for both comet and the restServer

+++ tests/test_dataset_broker.py
@@ -7,10 +7,12 @@ from future.builtins.disabled import *  # noqa  pylint: disable=W0401, W0614
 
 import pytest
 import os.path
+import re
 import time
-from subprocess import Popen
-from subprocess import call
+import tempfile
 import signal
+
+from subprocess import call, Popen
 
 producer_path = "../build/tests/boost/dataset_broker_producer"
 producer2_path = "../build/tests/boost/dataset_broker_producer2"
@@ -30,14 +32,36 @@ def test_produce_consume():
         print("Deactivated! Make sure the dataset_broker is at {}".format(broker_path))
         return
 
-    broker = Popen([broker_path])
-    time.sleep(2)
+    with tempfile.NamedTemporaryFile(mode="w") as f_out:
+        # Start comet with a random port
+        broker = Popen(
+            [broker_path, "-p", "0", "--recover", "False"], stdout=f_out, stderr=f_out
+        )
+        time.sleep(3)
 
-    try:
-        assert call([producer_path]) == 0
-        assert call([producer2_path]) == 0
-        assert call([consumer_path]) == 0
-    finally:
-        pid = broker.pid
-        os.kill(pid, signal.SIGINT)
-        broker.terminate()
+        # Find port in the log
+        regex = re.compile("Selected random port: ([0-9]+)$")
+        log = open(f_out.name, "r").read().split("\n")
+        port = None
+        for line in log:
+            print(line)
+            match = regex.search(line)
+            if match:
+                port = match.group(1)
+                print("Test found comet port in log: %s" % port)
+                break
+        if not match:
+            print("Could not find comet port in logs.")
+            exit(1)
+
+        try:
+            assert call([producer_path, "--", port]) == 0
+            assert call([producer2_path, "--", port]) == 0
+            assert call([consumer_path, "--", port]) == 0
+        finally:
+            pid = broker.pid
+            os.kill(pid, signal.SIGINT)
+            broker.terminate()
+            log = open(f_out.name, "r").read().split("\n")
+            for line in log:
+                print(line)
